{12,13,14,15,16,17,26}Migrations

+++ database/migrations/2022_01_11_133502_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->increments('id');
             $table->string('title');
             $table->string('description');
-            $table->integer('price');
+            $table->double('price');
             $table->string('fileType');
             $table->timestamps();
         });
